LPS-180879 Extract ScreenNavigationEntry methods from ServerSystemPropertiesScreenNavigationCategory to ServerSystemPropertiesScreenNavigationEntry, extends ServerSystemPropertiesScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
